Issue 4563 - Failure on s390x: 'Fails to split RDN  "o=pki-tomcat-CA" into components'
Bug Description:
SLAPI_OPERATION_TYPE is defined as unsigned long. In various places we
use int instead. This works fine on little-endian platforms, but fails
on big-endian, s390x in particular.

Fix Description:
Update all places where we use slapi_pblock_get with
SLAPI_OPERATION_TYPE to use unsigned long.

Fixes: https://github.com/389ds/389-ds-base/issues/4563

Reviewed by: ???

+++ ldap/servers/plugins/acl/acl.h
@@ -472,7 +472,7 @@ struct acl_pblock
 
     Slapi_Entry *aclpb_client_entry; /* A copy of client's entry */
     Slapi_PBlock *aclpb_pblock;      /* back to LDAP PBlock */
-    int aclpb_optype;                /* current optype from pb */
+    unsigned long aclpb_optype;                /* current optype from pb */
 
     /* Current entry/dn/attr evaluation info */
     Slapi_Entry *aclpb_curr_entry; /* current Entry being processed */
